+++ agg/src/agg_gsv_text.cpp
@@ -577,7 +577,7 @@ namespace agg
             m_text = m_chr;
             return;
         }
-        unsigned new_size = strlen(text) + 1;
+        unsigned new_size = (unsigned int)strlen(text) + 1;
         if(new_size > m_text_buf.size())
         {
             m_text_buf.resize(new_size);
